refactor: chane group bulk read p1
