+++ src/tachimawari/control/sdk/packet/protocol_2/group_sync_write.cpp
@@ -21,7 +21,7 @@
 #include <string>
 #include <vector>
 
-#include "tachimawari/control/sdk/utils/protocol_2/group_sync_write.hpp"
+#include "tachimawari/control/sdk/packet/protocol_2/group_sync_write.hpp"
 
 #include "tachimawari/joint/protocol_2/mx28_address.hpp"
